pyproject.toml
[tool.poetry]
name = "picture-combine"
version = "0.1.0"
description = ""
authors = ["onehandedpirate <plagueismkii@gmail.com>"]
readme = "README.md"
packages = [{include = "picture_combine"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
